chore: remove unnecessary cast usize
lint message: casting to the same type is unnecessary (`u64` -> `u64`)
help: for further information visit https://rust-lang.github.io/rust-clippy/rust-1.92.0/index.html#unnecessary_cast

+ disable `clippy::unnecessary_cast` for update_disk_usage
where the size of the input values depends on the platform

+++ src/db/fjall_adapter.rs
@@ -287,7 +287,7 @@ impl<'a> ITx for FjallTx<'a> {
 	}
 	fn len(&self, tree_idx: usize) -> TxOpResult<usize> {
 		let tree = self.get_tree(tree_idx)?;
-		Ok(self.tx.len(tree)? as usize)
+		Ok(self.tx.len(tree)?)
 	}
 
 	fn insert(&mut self, tree_idx: usize, key: &[u8], value: &[u8]) -> TxOpResult<()> {
